% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jab.adverse.reactions.R
\docType{package}
\name{jab.adverse.reactions}
\alias{jab.adverse.reactions}
\title{jab.adverse.reactions: Possible adverse events/reactions from the vaccinations/experimental gene therapies}
\description{
jab.adverse.reactions provides a single table that contains data about the
possible adverse events/reactions due to being injected with a
vaccine/experimental gene therapy shot.
}
