% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jab.adverse.reactions.R
\docType{data}
\name{jab_adverse_reactions}
\alias{jab_adverse_reactions}
\title{Vaccine/Experimental Gene Therapy Possible Adverse Events/Reactions}
\format{
A data.table data frame with 1324 rows and 6 variables:
\describe{
\item{Possible Vaccine Adverse Events}{Sorted, respective to the source, listing of possible adverse events/reactions resulting from the injection of a vaccine/experimental gene therapy jab}
\item{Source Title}{The title of the reference source}
\item{Source Author}{The author of the reference source}
\item{Source Date}{The date of the reference source}
\item{Source Additional Information}{Any other useful information from the citation of the reference source other than the URL}
\item{Source URL}{The URL of the reference source}
}
}
\usage{
jab_adverse_reactions
}
\description{
A table containing the possible adverse events/reactions due to being
injected with a vaccine/experimental gene therapy shot.
}
\examples{
jab_adverse_reactions
}
\references{
\enumerate{
   \item Worldwide Safety Pfizer, 5.3.6 Cumulative Analysis of Post-Authorization Adverse Event Reports of PF-07302048 (BNT162B2) Received Through 28-Feb-2021 Report. https://phmpt.org/wp-content/uploads/2021/11/5.3.6-postmarketing-experience.pdf.
   \item George Wilson, M.D. and Samuel B. Hadden, M.D., "Neuritis and Multiple Neuritis Following Serum Therapy", *JAMA*. 1932;98(2):123-125, January 9, 1932. https://jamanetwork.com/journals/jama/article-abstract/1153665. Also archived at https://archive.vn/Nw4nq.
   \item Steve Anderson, PhD, MPP, US Food and Drug Administration (FDA): Vaccines and Related Biological Products Advisory Committee October 22, 2020 Meeting Presentation, page 17. https://www.fda.gov/media/143557/download (Please Note: The document is no longer available from the original URL, but it is archived at https://web.archive.org/web/20201126033341/https://www.fda.gov/media/143557/download [Recovered with the Internet Archive: Wayback Machine])
}
}
